Compile the container in tests
During the compilation, a lot of things happen (validation, etc.). We should execute those things (i.e. compile the container) in tests to ensure that we don't miss some edge cases. For example there is a bug in `factory` (see #28) that we missed because of that.

This commit refactors the tests so that the container is always compiled.

Due to this commit, tests are failing (i.e. it reproduces #28). Tests will be fixed in another commit for more readability.