Fix #28: factories must have a class name
```diff
 return [
-    'newsletter_manager' => factory([NewsletterManager::class, 'create'])
+    'newsletter_manager' => factory([NewsletterManager::class, 'create'], NewsletterManager::class)
         ->arguments('foo', 'bar'),
 ];
```

This fixes the exception:

> Please add the class to service "config" even if it is constructed by a factory since we might need to add method calls based on compile-time checks.

Symfony documentation:

> When using a factory to create services, the value chosen for the class option has no effect on the resulting service. The actual class name only depends on the object that is returned by the factory. However, the configured class name may be used by compiler passes and therefore should be set to a sensible value.

To keep DX good, the class name parameter can be skipped: the service ID will be used as the class name. This can be used when using class names as service IDs, for example:

```php
return [
    // you can write:
    NewsletterManager::class => factory([NewsletterManager::class, 'create']),
    // instead of:
    NewsletterManager::class => factory([NewsletterManager::class, 'create'], NewsletterManager::class),
];
```

This is based on #29 which refactors tests to cover #28 (which is why there are no tests changes in this commit: tests are already failing because this commit is based on #29).